Changed Calculator.cpp, Calculator.h, Tester.cpp, Tester.h.
Main changes were to the "power" and "exponent" functions, had to add logic that allowed negative exponents to be implemented correctly.

Minor changes to "power" and "exponents" return types and parameters. Changes ints to floats in certain places in order to avoid truncation. Did this for the "tester" parameters as well.

Also added the .sln file to the repository.

Will be much more detailed and specific the next time I make changes to a Git repository.